+++ src/app/math/impl/mathUtilOwn.ts
@@ -8,16 +8,19 @@ export class MathUtilOwn implements MathUtil {
 
   mean(numbers: number[]): number {
     const sum = this.sum(numbers);
+
     return sum / numbers.length;
   }
 
   median(numbers: number[]): number {
-    if (numbers.length === 0) return 0;
+    if (0 === numbers.length) {
+      return 0;
+    }
 
     const half = Math.floor(numbers.length / 2);
-    const numbersOrdered = numbers.sort();  
-    
-    if (numbers.length % 2 === 0) {
+    const numbersOrdered = numbers.sort();
+
+    if (0 === numbers.length % 2) {
       return (numbersOrdered[half - 1] + numbersOrdered[half]) / 2.0;
     }
 
@@ -29,23 +32,23 @@ export class MathUtilOwn implements MathUtil {
   }
 
   mode(numbers: number[]): number[] {
-    const modes: number[] = []
-    let count = {};
+    const modes: number[] = [];
+    const count = {};
     let maxRepeat = 0;
- 
+
     numbers.forEach((number: number) => {
       count[number] = (count[number] || 0) + 1;
       if (count[number] > maxRepeat) {
         maxRepeat = count[number];
       }
-    })
+    });
 
     for (const number in count) {
       if (count[number] === maxRepeat) {
-          modes.push(Number(number));
+        modes.push(Number(number));
       }
     }
- 
+
     return modes;
   }