+++ src/app/math/util/math.function.ts
@@ -3,7 +3,7 @@ import { MathUtil } from './math.util.interface';
 export class MathFunctions {
 
   constructor(private mathUtilImpl: MathUtil) { }
-  
+
   max(numbers: number[]): number {
     return this.mathUtilImpl.max(numbers);
   }